add city blueprint
