+++ web_travel/crud.py
@@ -1,6 +1,8 @@
-from web_travel.models import db, User, City, Place
-from web_travel.country.models import Country
 import json
+
+from web_travel.models import db, User, Place
+from web_travel.country.models import Country
+from web_travel.city.models import City
 
 
 def save_user(username, password, email):
