+++ web_travel/templates/city/add.html
@@ -1,0 +1,27 @@
+{% extends 'base.html'%}
+
+{% block content %}
+    {% with messages = get_flashed_messages() %}
+        {% if messages %}
+        <div class="alert alert-danger" role="alert">
+            {% for message in messages %}
+                {{ message }}<br>
+            {% endfor %}
+        </div>
+        {% endif %}
+    {% endwith %}
+    <div class="row">
+        <div class="col-8">
+            <form action=""" method='POST'>
+                {{ form.hidden_tag() }}
+                <div class="form-group">
+                    {{ form.country_name.label }}
+                    {{ form.country_name() }}
+                </div>
+                {{ form.submit() }}
+            </form>
+        </div>
+        <div class="col-4"></div>
+    </div>
+
+{% endblock %}
